CA-210762: Linux bridge: Always update bond properties
Bond properties such as the bond mode were not updated if the bond already
existed. This caused XenAPI commands such as Bond.set_mode to have no immediate
effect (changes would be implemented only after re-plugging or rebooting).

Also make sure that bond-manipulation functions in Network_utils are idempotent,
as the high-level function call is supposed to be idempotent.

Signed-off-by: Rob Hoes <rob.hoes@citrix.com>