CP-23601 Inject IGMP snooping query message when IGMP snooping toggle
In this change, when setting IGMP snooping toggle mcast_snooping_enable in OVS,
if the current setting is off but new setting is on, then a script will be called to inject a IGMP query into
the bridge (each vif interface), for quicker adaption to IGMP for applications in VMs.
A tunable configuration field igmp-query-maxresp-time is added, for maximum responsible time of the injected query, default to be at 5000ms.

Signed-off-by: Yarsin He <yarsin.he@citrix.com>