CP-26923: Simplify Networkd logic by adding comments in modprobe config file.
For those drivers don't support sysfs interface. We add comments in
modprobe config file to identify the VF param and max_vfs. When we
enable SR-IOV, we first check the existence of VF param to determine
using modprobe or sysfs. If it is a modprobe driver, we write the
configuration into the config file.

- Add a function to determine if the VF param is an array or not.
- Read config from comments in driver config file.
- Move config_sriov to Modprobe module.
- Refine get_capabilities
- Add modinfo module

Signed-off-by: Wei Xie <wei.xie@citrix.com>