CA-236855: VM.clean_shutdown task not completing on slave
From some historical test, it is found xapi (events_watch thread, for xenopsd events) is stuck some where some time. Further analysis found it is due to some earlier xapi->xcp-networkd call did not get response hence some mutex is not release so later use of the mutex is blocked. The reason the xapi->xcp-networkd is not responsed is further due to xcp-networkd to PVS proxy is not responded. After discussion with Rob and Jon, it is decided to add a time guard for xcp-networkd for the call to PVSproxy.

The implementation added a timebox call, in which another thread is created to do the actual RPC call, and the original thread will wait for some configurable timeout seconds, and proceed if the call did not ended.

Signed-off-by: Yarsin He <yarsin.he@citrix.com>