CA-291495: Fall back to Ip.link_set_mtu if OVS.set_mtu fails
OVS.set_mtu is a new function (since ed2151f8), which requires a newer version
of the OVS that supports the mtu_request feature. If the installed OVS does not
have this feature, then the function will raise an exception. For compatibility
reasons, we now catch this exception and fall back to Ip.link_set_mtu.

Signed-off-by: Rob Hoes <rob.hoes@citrix.com>