CA-291197: Toolstack no longer ignores junk at the end of JSON-RPC responses from pvsproxy
This is resulted by CA-236855, in that the read strategy is changed to assume all good
data will come from PVSproxy side. However, in testing, it is found a bug in PVSproxy,
which was not fatal locally, but resulted issue in overall scenario. It is decided to
enhance the solution a little bit based on previous change.

The updates as to integrate some part in the old byte by byte read, where a function
is used to prescreen the data read from PVSproxy, and tailing junk is discarded, so
the only json part is further processed. Here we most copied the original function,
but added:
 1. to get input chars from a string;
 2. to tolerant heading spaces/junk (in screening);

Signed-off-by: YarsinCitrix <yarsin.he@citrix.com>